test: move `test_filter_order_handler.py` to `test` dir #265
